Update charting functions.
git-svn-id: svn://biometris-nas.wurnet.nl/AMIGA@308 3ae013c2-4042-2242-b34d-985be216b4aa

+++ AmigaPowerAnalysis/Biometris/Statistics/Histograms/HistogramBinUtilities.cs
@@ -62,7 +62,7 @@ namespace Biometris.Statistics {
         /// </summary>
         /// <param name="bins"></param>
         /// <returns></returns>
-        public static int GetTotalFrequency(this IEnumerable<HistogramBin> bins) {
+        public static double GetTotalFrequency(this IEnumerable<HistogramBin> bins) {
             return bins.Sum(b => b.Frequency);
         }
